Entrypoints Integration Test (Pooling)
