+++ tests/entrypoints/pooling/correctness/test_mteb_embed.py
