+++ tests/entrypoints/pooling/correctness/test_mteb_score.py
