+++ tests/entrypoints/pooling/llm/test_classify.py
@@ -6,12 +6,13 @@ import weakref
 import pytest
 import torch
 
+from tests.models.utils import softmax
 from vllm import LLM, PoolingParams
 from vllm.distributed import cleanup_dist_env_and_memory
 
-from ...models.utils import softmax
+MODEL_NAME = "jason9693/Qwen2.5-1.5B-apeach"
 
-MODEL_NAME = "tomaarsen/Qwen3-Reranker-0.6B-seq-cls"
+prompts = ["The chef prepared a delicious meal."]
 
 
 @pytest.fixture(scope="module")
@@ -36,15 +37,11 @@ def llm():
 def test_pooling_params(llm: LLM):
 
     def get_outputs(activation):
-        text_1 = "What is the capital of France?"
-        text_2 = "The capital of France is Paris."
-
-        outputs = llm.score(
-            text_1,
-            text_2,
+        outputs = llm.classify(
+            prompts,
             pooling_params=PoolingParams(activation=activation),
             use_tqdm=False)
-        return torch.tensor([x.outputs.score for x in outputs])
+        return torch.tensor([x.outputs.probs for x in outputs])
 
     default = get_outputs(activation=None)
     w_activation = get_outputs(activation=True)
@@ -58,3 +55,15 @@ def test_pooling_params(llm: LLM):
     assert torch.allclose(
         softmax(wo_activation), w_activation, atol=1e-2
     ), "w_activation should be close to activation(wo_activation)."
+
+
+def test_encode_api(llm: LLM):
+    err_msg = "pooling_task must be one of.+"
+    with pytest.raises(ValueError, match=err_msg):
+        llm.encode(prompts, use_tqdm=False)
+
+
+def test_score_api(llm: LLM):
+    err_msg = "Score API is only enabled for num_labels == 1."
+    with pytest.raises(ValueError, match=err_msg):
+        llm.score("ping", "pong", use_tqdm=False)
