+++ tests/entrypoints/pooling/llm/test_embedding.py
