+++ tests/entrypoints/pooling/llm/test_encode.py
