+++ tests/entrypoints/pooling/openai/test_embedding_long_text.py
@@ -14,9 +14,8 @@ import openai
 import pytest
 import pytest_asyncio
 
+from tests.utils import RemoteOpenAIServer
 from vllm.entrypoints.openai.protocol import EmbeddingResponse
-
-from ...utils import RemoteOpenAIServer
 
 
 def _generate_random_text(word_count: int) -> str:
