+++ tests/entrypoints/pooling/openai/test_truncation.py
